Added update password btn on profile pages to display password fields on click and hide password fields on second click, also added confirmation dialog on click of delete btns

+++ frontend/src/ShelterManagement.js
@@ -30,9 +30,11 @@ const ShelterManagement = () => {
 
     const delete_pet = (id) => {
         // console.log(id + ' Delete clicked!');
-        fetch(fetchURL + '/pets/' + id, { method: 'DELETE'}).then(() => {
-            setPetFromAPI(petsFromAPI.filter((pet) => pet.id !== id));
-        });
+        if (window.confirm('Are you sure you wish to delete this pet?')){
+            fetch(fetchURL + '/pets/' + id, { method: 'DELETE'}).then(() => {
+                setPetFromAPI(petsFromAPI.filter((pet) => pet.id !== id));
+            });
+        }
     };
 
     return (
